refactor: remove input port directory for category
