+++ internal/adapters/primary/graph/resolver.go
@@ -4,11 +4,11 @@
 package graph
 
 import (
-	"github.com/lechitz/AionApi/internal/core/ports/input/graphql"
+	"github.com/lechitz/AionApi/internal/core/ports/input"
 	"github.com/lechitz/AionApi/internal/core/ports/output/logger"
 )
 
 type Resolver struct {
-	CategoryService graphql.CategoryService
+	CategoryService input.CategoryService
 	Logger          logger.Logger
 }
